Back port --self-update --major (#92)
[Feature]
- Restrict major version updates with --self-update unless --major is specified

To prepare for the upcoming major update, I have modified the --self-update behavior. Now, updating to a new major HeavyScript version requires the explicit use of the --major flag.

The new CLI is under development and can be found on the "args" branch. It will remain there until it's fully developed and tested.