Allow non-root & Remove Patches Function (#140)
[Features]
- Allow one-liner-scipt install for non-root users closes #118 #94 
- Asks users to run certain root-requiring functions as root/sudo
- Create symbolic link in /usr/local/bin for heavyscript wrapper, for the sudo environment


[Notes]
We have removed the patches function along with its corresponding menu option. This change comes in response to iXsystems discontinuing support for all HeavyScript users, irrespective of their use case. It appears that any issues raised on systems with HeavyScript installed are being closed by iXsystems, citing "HeavyScript modifies core files" as the reason.

We suspect that the discontinuation of support might be due to the patches function, which was operative only on a specific version of TrueNAS SCALE. This function was designed to implement a patch that was scheduled to be included in the subsequent release of TrueNAS SCALE. It's important to note that HeavyScript included a version check to ensure the patches function would only run on the EXACT version it was intended for.

The specific version of TrueNAS SCALE, 22.12.0, was released nearly a year ago. Given that HeavyScript had a built-in check to prevent the patches function from running on any other version, it is perplexing why iXsystems has only now chosen to withdraw support, especially when this functionality has been irrelevant for almost a year.