Fix link errors (#162)
Resolves #160

Due to recent policy changes in Dragonfish, modifications to certain files and directories have been restricted. This has necessitated a rethink in how we manage the visibility of paths between the user invoking sudo and the sudo environment itself.

This update renames the `ensure_symlink` function to `ensure_sudoers` and modifies its functionality to better align with the new restrictions. Now, instead of creating symlinks, the function appends two lines to the sudoers file. These lines ensure that the `PATH` for users invoking sudo includes `$HOME/bin`, making it visible and accessible within the sudo environment.

Changes:
- Renamed `ensure_symlink` to `ensure_sudoers`.
- Modified function to append `$HOME/bin` to the sudoer's `PATH`.

This adjustment simplifies the configuration and enhances compliance with the updated Dragonfish policies.